requirements.txt
streamlit
pandas
spacy
pdfminer.six
Pillow
docx2txt
sentence-transformers
